requirements.txt
# PulseAI — Python Dependencies

# News Aggregation
feedparser==6.0.11
requests==2.32.3

# AI / LLM
langchain==0.3.15
langchain-groq==0.2.3
groq==0.13.0

# Database
supabase==2.10.0
pydantic>=2.10.0

# Editorial Dashboard
streamlit==1.41.1

# Scheduling
apscheduler==3.10.4

# FastAPI Backend (Phase 2)
fastapi==0.115.12
uvicorn==0.34.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
